requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.2.2
scikit-learn>=1.5.1
xgboost>=2.1.1
pytest>=8.3.0
